mobile: refactor add device, bug fixes
